Delay purging of sync group data for a specified time after switching (#2023)
* Add local shared preference for tracking the time at which facility sync group was last switched

* Add `delayPurgeAfterSwitchFor` property to `PurgeAfterSync`

* Provide the current UTC clock as a dependency to `PurgeAfterSync`

* Delay purging of data from other sync groups

* Use a remote config field for fetching the facility sync group purge delay time

* Change method in `FacilityRepository` to set the current facility to a synchronous call

* Remove deprecated trampoline schedulers provider usage

* Add effect to load the current facility

* Add the current facility to the model

* When the confirm facility change sheet is opened, load the current facility

* Track the current facility in the model

* Pass the new facility to the event after switching facilities

* Add facility sync group switched at time preference to the effect handler

* Add effect to update the facility sync group switched at time

* Update the facility sync group switched at timestamp

* Update CHANGELOG

+++ app/src/main/java/org/simple/clinic/sync/PurgeOnSync.kt
@@ -1,18 +1,56 @@
 package org.simple.clinic.sync
 
 import androidx.annotation.WorkerThread
+import com.f2prateek.rx.preferences2.Preference
 import org.simple.clinic.AppDatabase
 import org.simple.clinic.facility.Facility
+import org.simple.clinic.main.TypedPreference
+import org.simple.clinic.main.TypedPreference.Type.FacilitySyncGroupSwitchedAt
+import org.simple.clinic.remoteconfig.ConfigReader
+import org.simple.clinic.util.Optional
+import org.simple.clinic.util.UtcClock
+import java.time.Duration
+import java.time.Instant
 import javax.inject.Inject
 import javax.inject.Provider
 
-class PurgeOnSync @Inject constructor(
+class PurgeOnSync(
     private val currentFacility: Provider<Facility>,
-    private val appDatabase: AppDatabase
+    private val appDatabase: AppDatabase,
+    private val facilitySyncGroupSwitchedAt: Preference<Optional<Instant>>,
+    private val delayPurgeAfterSwitchFor: Duration,
+    private val clock: UtcClock
 ) {
+
+  @Inject
+  constructor(
+      currentFacility: Provider<Facility>,
+      appDatabase: AppDatabase,
+      @TypedPreference(FacilitySyncGroupSwitchedAt) facilitySyncGroupSwitchedAt: Preference<Optional<Instant>>,
+      clock: UtcClock,
+      remoteConfig: ConfigReader
+  ) : this(
+      currentFacility = currentFacility,
+      appDatabase = appDatabase,
+      facilitySyncGroupSwitchedAt = facilitySyncGroupSwitchedAt,
+      delayPurgeAfterSwitchFor = Duration.parse(remoteConfig.string("delay_purge_after_facility_sync_group_switch_duration", "P1D")),
+      clock = clock
+  )
 
   @WorkerThread
   fun purgeUnusedData() {
-    appDatabase.deletePatientsNotInFacilitySyncGroup(currentFacility.get())
+    val facilitySyncSwitchedAtInstant = facilitySyncGroupSwitchedAt.get()
+
+    val shouldPurgeData = !facilitySyncSwitchedAtInstant.isPresent() || isSafeToPurgeData(facilitySyncSwitchedAtInstant.get())
+
+    if(shouldPurgeData) appDatabase.deletePatientsNotInFacilitySyncGroup(currentFacility.get())
+  }
+
+  private fun isSafeToPurgeData(
+      switchedAt: Instant
+  ): Boolean {
+    val now = Instant.now(clock)
+
+    return Duration.between(switchedAt, now) > delayPurgeAfterSwitchFor
   }
 }
